django_app/requirements.txt
Django==4.2.0
djangorestframework==3.14.0
django-cors-headers==4.1.0
Pillow==10.0.0
django-crispy-forms==2.0
crispy-bootstrap5==0.7
django-filter==23.3
